revised feeder parts for testing
These parts are intended for testing an upgrade of the feeder housing.

The parts are redesigned so that the contact areas between the filament
and the feeder housing are eliminated. The outsides of the housing parts
are flat for easy printing.

In order to test this revision you will need a metal ring that can be
positioned in the filament inlet.